Fix type

+++ src/UnfoldedCircle.Electrolux/Http/Models/ApplianceReportedProperty.cs
@@ -5,7 +5,7 @@ public sealed record ApplianceReportedProperty(
     [property: JsonPropertyName("Fanspeed")] sbyte FanSpeed,
     [property: JsonPropertyName("TVOC")] ushort Tvoc,
     [property: JsonPropertyName("CO2")] ushort Co2,
-    [property: JsonPropertyName("Temp")] short Temperature,
+    [property: JsonPropertyName("Temp")] sbyte Temperature,
     [property: JsonPropertyName("Humidity")] sbyte Humidity,
     [property: JsonPropertyName("PM1")] ushort Pm1,
     [property: JsonPropertyName("PM2_5")] ushort Pm25,
